+++ CHANGES.md
@@ -5,6 +5,7 @@ Release Notes.
 9.5.0
 ------------------
 
+* Add the virtual thread executor plugin
 
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/236?closed=1)